// src/features/funding/types.ts
export enum DepositQuoteStatus {
  Error = 'error',
  InsufficientBalance = 'insufficientBalance',
  InsufficientGas = 'insufficientGas',
  Pending = 'pending',
  Success = 'success',
  ZeroAmountError = 'errorZeroAmount',
}